Modifying nck function on main.py to return int istead of float

+++ main.py
@@ -4,4 +4,4 @@ def nck(n , k):
     """N Chose K function to get the number of K combinations of N elements given by the binomial.
     ARGS: n, k where n is the number of elements and k is the number of combinations we need to get from n
     """
-    return F(n) / (F(k) * F(n-k))
+    return int(F(n) / (F(k) * F(n-k)))
